+++ Backend/application/resources/commissioner/commissioner_category_delete_resource.py
@@ -20,6 +20,9 @@ def commissioner_delete_category(
     category = db.get(Department, category_id)
     if not category:
         raise HTTPException(status_code=404, detail="Category not found")
+        
+    if category.users_in_dept:
+        raise HTTPException(status_code=400, detail="Cannot delete department. There are officers present in this department.")
 
     db.delete(category)
     db.commit()
