+++ Backend/application/resources/commissioner/commissioner_facility_delete_resource.py
@@ -4,6 +4,8 @@ from application.extensions.db_extn import get_db
 from application.helpers.models import User, Facility
 from application.middlewares.init_jwt import get_current_user_id
 
+from datetime import date
+
 router = APIRouter()
 
 
@@ -20,6 +22,10 @@ def commissioner_delete_facility(
     facility = db.get(Facility, facility_id)
     if not facility:
         raise HTTPException(status_code=404, detail="Facility not found")
+        
+    has_future_bookings = any(booking.booked_date >= date.today() for booking in facility.bookings)
+    if has_future_bookings:
+        raise HTTPException(status_code=400, detail="Cannot delete facility. There are future bookings associated with it.")
 
     db.delete(facility)
     db.commit()
